backend/requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Audio recognition (AcoustID — free & open source; requires fpcalc on PATH)
pyacoustid==1.3.0            # AcoustID API client
httpx==0.27.0                # Spotify + AcoustID HTTP calls

# Audio processing (mel spectrogram for CnnMusicEncoder)
librosa==0.10.2
soundfile==0.12.1

# ML inference
numpy>=2.0.0                 # 2.x required for Python 3.13 prebuilt wheels
faiss-cpu>=1.8.0             # swap for faiss-gpu if running on the same GPU host
onnxruntime>=1.18.0          # runs the exported CnnMusicEncoder ONNX model

# Utilities
python-dotenv==1.0.1
pydantic>=2.10.0             # 2.10+ has Python 3.13 prebuilt wheels
pydantic-settings>=2.6.0
cachetools>=5.3.3
